Stop PK API client from modifying root logger
Previously, importing the placekey module would cause it to immediately
suppress all but logging.ERROR logs in any code that it is imported
into. This is somewhat unexpected since you'd expect that importing a
module doesn't have side-effects on the rest of your program.

This makes it so that the logging settings in this module apply to this
module.